feat: Specification pattern

+++ common-domain/src/main/java/com/manager/shared/domain/model/dto/PagedResponse.java
@@ -5,12 +5,12 @@ import java.util.List;
 public record PagedResponse<T>(
         List<T> items,
         int currentPage,
-        int totalItems,
+        long totalItems,
         int totalPages,
         boolean hasNext
 ) {
 
-    public static <T> PagedResponse<T> of(List<T> items, int page, int size, int total) {
+    public static <T> PagedResponse<T> of(List<T> items, int page, int size, long total) {
         int totalPages = (int) Math.ceil((double) total / size);
         boolean hasNext = page < totalPages;
         return new PagedResponse<T>(items, page, total, totalPages, hasNext);
